delegate Item component from Items to components/Item
